+++ lib/chat_message_updater.rb
@@ -25,9 +25,9 @@ class DiscourseChat::ChatMessageUpdater
       @chat_message.save!
       save_revision!
       update_uploads!
-      update_mention_notifications
       ChatPublisher.publish_edit!(@chat_channel, @chat_message)
       Jobs.enqueue(:process_chat_message, { chat_message_id: @chat_message.id })
+      Jobs.enqueue(:send_chat_notifications, { type: :edit, chat_message_id: @chat_message.id })
     rescue => error
       puts error.inspect
       @error = error
@@ -57,39 +57,5 @@ class DiscourseChat::ChatMessageUpdater
 
   def save_revision!
     @chat_message.revisions.create!(old_message: @old_message_content, new_message: @chat_message.message)
-  end
-
-  def update_mention_notifications
-    existing_notifications = Notification
-      .where(notification_type: Notification.types[:chat_mention])
-      .where("data LIKE ?", "%\"chat_message_id\":#{@chat_message.id}%")
-
-    already_notified_user_ids = existing_notifications.map(&:user_id)
-    mentioned_user_ids = DiscourseChat::ChatMessageCreator.
-      mentioned_users(chat_message: @chat_message, creator: @user)
-      .map(&:id)
-
-    needs_deletion = already_notified_user_ids - mentioned_user_ids
-    needs_deletion.each do |user_id|
-      notification = existing_notifications.detect { |n| n.user_id == user_id }
-      notification.destroy!
-    end
-
-    needs_notification_ids = mentioned_user_ids - already_notified_user_ids
-    return unless needs_notification_ids.present?
-
-    needs_notification = User
-      .includes(:do_not_disturb_timings, :user_chat_channel_memberships)
-      .where(id: needs_notification_ids)
-    needs_notification.each { |target_user|
-      if Guardian.new(target_user).can_see_chat_channel?(@chat_channel)
-        DiscourseChat::ChatMessageCreator.create_mention_notification(
-          creator_username: @user.username,
-          mentioned_user: target_user,
-          chat_channel: @chat_channel,
-          chat_message: @chat_message,
-        )
-      end
-    }
   end
 end
