--- lib/chat_notifier.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class DiscourseChat::ChatNotifier
-
-end
